Moved Models to shared maven package to make it easier to use them in the backend
