+++ .idea/encodings.xml
@@ -3,5 +3,7 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/client/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/client/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/models/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/models/src/main/resources" charset="UTF-8" />
   </component>
 </project>
