+++ client/src/main/java/dev/icefish/tourplanner/client/controllers/TourDetailViewController.java
@@ -1,6 +1,7 @@
 package dev.icefish.tourplanner.client.controllers;
 
-import dev.icefish.tourplanner.client.model.Tour;
+import dev.icefish.tourplanner.models.Tour;
+import dev.icefish.tourplanner.models.TourLog;
 import dev.icefish.tourplanner.client.utils.WindowUtils;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
