+++ client/src/main/java/dev/icefish/tourplanner/client/services/MapService.java
@@ -1,0 +1,4 @@
+package dev.icefish.tourplanner.client.services;
+
+public class MapService {
+}
