+++ client/src/main/java/module-info.java
@@ -3,10 +3,11 @@ module dev.icefish.tourplanner.client {
     requires javafx.fxml;
     requires javafx.graphics;
     requires java.sql;
+    requires dev.icefish.tourplanner.models;
+
 
     opens dev.icefish.tourplanner.client to javafx.fxml;
     exports dev.icefish.tourplanner.client;
     exports dev.icefish.tourplanner.client.controllers;
-    exports dev.icefish.tourplanner.client.model;
     opens dev.icefish.tourplanner.client.controllers to javafx.fxml;
 }
