+++ models/pom.xml
@@ -1,20 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>dev.icefish.tourplanner</groupId>
+        <artifactId>tourplanner</artifactId>
+        <version>1.0</version>
+    </parent>
 
-    <groupId>dev.icefish.tourplanner</groupId>
-    <artifactId>tourplanner</artifactId>
-    <version>1.0</version>
-    <packaging>pom</packaging>
+    <artifactId>models</artifactId>
 
-    <modules>
-        <module>client</module>
-        <module>server</module>
-        <module>models</module>
-    </modules>
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+
 </project>
