+++ models/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module dev.icefish.tourplanner.models {
+    requires java.sql;
+    exports dev.icefish.tourplanner.models;
+}
