+++ server/pom.xml
@@ -23,6 +23,11 @@
             <artifactId>gson</artifactId>
             <version>2.8.9</version>
         </dependency>
+        <dependency>
+            <groupId>dev.icefish.tourplanner</groupId>
+            <artifactId>models</artifactId>
+            <version>1.0</version>
+        </dependency>
     </dependencies>
 
     <build>
