+++ server/src/main/java/dev/icefish/tourplanner/server/repository/TourLogRepository.java
@@ -1,0 +1,4 @@
+package dev.icefish.tourplanner.server.repository;
+
+public class TourLogRepository {
+}
